# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Clothing Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  imageUrl?: string;
}