# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Clothing Store API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  productId: number;
  quantity: number;
  price: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  productName: string;
}