# lib/api-zod/src/generated/types/outfitRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Clothing Store API
 * OpenAPI spec version: 0.1.0
 */

export interface OutfitRequest {
  productId: number;
  occasion?: string;
}