# lib/api-zod/src/generated/types/outfitSuggestions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Clothing Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface OutfitSuggestions {
  outfitDescription: string;
  suggestions: Product[];
}